WIP Resource Browser in forms
