+++ modules/frontend/frontend/src/components/scope/visualizers/Resource.tsx
@@ -1,0 +1,126 @@
+/* eslint-disable no-console */
+import * as React from "react";
+import {
+  FormProps,
+  SingleElementVisualizerProps,
+  visualizersForEventType,
+} from "../../../registry/visualization";
+import {
+  StandardComponentOptions,
+  StandardComponent,
+} from "./StandardComponent";
+import { Card } from "./Card";
+import { Resource } from "@farm-ng/genproto-perception/farm_ng/core/resource";
+import { useFormState } from "../../../hooks/useFormState";
+import Form from "./Form";
+import { useFetchResource } from "../../../hooks/useFetchResource";
+import { parseResourceContentType } from "../../../utils/protoConversions";
+import styles from "./Resource.module.scss";
+import { Button, Modal } from "react-bootstrap";
+import { Icon } from "../../Icon";
+import { useState } from "react";
+import { BlobstoreBrowser } from "../../BlobstoreBrowser";
+
+const ResourceForm: React.FC<FormProps<Resource>> = (props) => {
+  const [value, setValue] = useFormState(props);
+  const [showModal, setShowModal] = useState(false);
+
+  const handleCloseModal = () => setShowModal(false);
+  const handleShowModal = () => setShowModal(true);
+  const handleFileSelected = (path: string) =>
+    setValue((v) => ({
+      ...v,
+      path,
+    }));
+
+  // A form,
+  // with two text fields for path and content_type (eventually support data too)
+  // and a button next to path that launches a modal, allowing for selecting of a path
+  // modal can select, preview, confirm or cancel.
+  return (
+    <>
+      <Form.Group
+        label="Resource Path"
+        value={value.path}
+        type="text"
+        description={value.contentType}
+        append={
+          <Button variant="outline-secondary" onClick={handleShowModal}>
+            <Icon id="folder2-open" />
+          </Button>
+        }
+        onChange={(e) => {
+          const path = e.target.value;
+          setValue((v) => ({
+            ...v,
+            path,
+          }));
+        }}
+      />
+
+      <Modal
+        show={showModal}
+        onHide={handleCloseModal}
+        dialogClassName={styles.modal}
+      >
+        <Modal.Header closeButton>
+          <Modal.Title>Select a file</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <BlobstoreBrowser
+            onFileSelected={handleFileSelected}
+            editingEnabled={false}
+            openInNewTabEnabled={false}
+            updateUrlEnabled={false}
+          ></BlobstoreBrowser>
+        </Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={handleCloseModal}>
+            Close
+          </Button>
+          <Button variant="primary" onClick={handleCloseModal}>
+            Select
+          </Button>
+        </Modal.Footer>
+      </Modal>
+    </>
+  );
+};
+
+const ResourceElement: React.FC<SingleElementVisualizerProps<Resource>> = (
+  props
+) => {
+  const {
+    value: [timestamp, value],
+    resources,
+  } = props;
+
+  const [, typeUrl] = parseResourceContentType(value);
+  const visualizer = visualizersForEventType(typeUrl)[0];
+  if (!visualizer || !visualizer.Element) {
+    return null;
+  }
+
+  const data = useFetchResource(value, resources);
+  if (!data) {
+    return null;
+  }
+
+  return (
+    <Card timestamp={timestamp} json={value}>
+      {React.createElement(visualizer.Element, {
+        ...props,
+        value: [timestamp, data],
+      })}
+    </Card>
+  );
+};
+
+export const ResourceVisualizer = {
+  id: "Resource",
+  types: ["type.googleapis.com/farm_ng.core.Resource"],
+  options: StandardComponentOptions,
+  Component: StandardComponent(ResourceElement),
+  Element: ResourceElement,
+  Form: ResourceForm,
+};
